Refactor ExistsOn to match new structure
Also fixed a few minor issues with the behavior of lists in the annotation.
