+++ Common/src/transform/groovy/io/github/groovymc/cgl/impl/transform/environment/AbstractExistsOnASTTransformer.groovy
@@ -3,21 +3,20 @@
  * SPDX-License-Identifier: LGPL-3.0-or-later
  */
 
-package io.github.groovymc.cgl.transform.environment
+package io.github.groovymc.cgl.impl.transform.environment
 
 import groovy.transform.CompileStatic
 import io.github.groovymc.cgl.api.environment.Platform
+import io.github.groovymc.cgl.api.transform.environment.ExistsOn
+import io.github.groovymc.cgl.impl.transform.TransformUtils
 import org.codehaus.groovy.ast.ASTNode
 import org.codehaus.groovy.ast.AnnotatedNode
 import org.codehaus.groovy.ast.AnnotationNode
 import org.codehaus.groovy.ast.ClassNode
-import org.codehaus.groovy.ast.expr.*
 import org.codehaus.groovy.control.CompilePhase
 import org.codehaus.groovy.control.SourceUnit
 import org.codehaus.groovy.transform.AbstractASTTransformation
 import org.jetbrains.annotations.ApiStatus
-
-import java.util.function.Function
 
 import static org.codehaus.groovy.ast.ClassHelper.makeWithoutCaching
 
@@ -38,39 +37,10 @@ abstract class AbstractExistsOnASTTransformer extends AbstractASTTransformation 
         if (MY_TYPE != anno.getClassNode()) return
 
         //noinspection UnnecessaryQualifiedReference
-        List<Platform> platforms = AbstractExistsOnASTTransformer.<Platform>getMemberValues(anno, 'value', Platform::valueOf)
+        List<Platform> platforms = TransformUtils.<Platform>getMemberValues(anno, 'value', Platform::valueOf)
 
         PROCESSOR?.process(parent, platforms, source.AST, phase)
     }
 
     abstract CompilePhase getPhase();
-
-    static <T extends Enum> List<T> getMemberValues(AnnotationNode anno, String name, Function<String,T> getter) {
-        Expression expr = anno.getMember(name)
-        if (expr == null) {
-            return []
-        }
-        if (expr instanceof ListExpression) {
-            final ListExpression listExpression = (ListExpression) expr
-            List<T> list = new ArrayList<>()
-            for (Expression itemExpr : listExpression.getExpressions()) {
-                if (itemExpr instanceof ConstantExpression) {
-                    T value = parseSingleExpr(itemExpr, getter)
-                    if (value != null) list.add(value)
-                }
-            }
-            return list
-        }
-        T single = parseSingleExpr(expr, getter)
-        return single==null?([]):([single])
-    }
-
-    static <T extends Enum> T parseSingleExpr(Expression itemExpr, Function<String,T> getter) {
-        if (itemExpr instanceof VariableExpression) {
-            return getter.apply(itemExpr.text)
-        } else if (itemExpr instanceof PropertyExpression) {
-            return getter.apply(itemExpr.propertyAsString)
-        }
-        return null
-    }
 }
