+++ Quilt/src/transform/groovy/io/github/groovymc/cgl/impl/transform/quilt/ExistsOnProcessorImpl.groovy
@@ -3,50 +3,39 @@
  * SPDX-License-Identifier: LGPL-3.0-or-later
  */
 
-package io.github.groovymc.cgl.forge.transform.environment
+package io.github.groovymc.cgl.impl.transform.quilt
 
 import groovy.transform.CompileStatic
 import io.github.groovymc.cgl.api.environment.Loader
 import io.github.groovymc.cgl.api.environment.Platform
 import io.github.groovymc.cgl.api.environment.Side
-import io.github.groovymc.cgl.transform.environment.ExistsOnProcessor
-import net.minecraftforge.api.distmarker.Dist
-import net.minecraftforge.api.distmarker.OnlyIn
+import io.github.groovymc.cgl.impl.transform.environment.ExistsOnProcessor
 import org.codehaus.groovy.ast.*
-import org.codehaus.groovy.ast.expr.ClassExpression
-import org.codehaus.groovy.ast.expr.ConstantExpression
-import org.codehaus.groovy.ast.expr.PropertyExpression
 import org.codehaus.groovy.control.CompilePhase
 import org.jetbrains.annotations.ApiStatus
+import org.quiltmc.loader.api.minecraft.ClientOnly
+import org.quiltmc.loader.api.minecraft.DedicatedServerOnly
 
 import static org.codehaus.groovy.ast.ClassHelper.makeWithoutCaching
 
 @CompileStatic
 @ApiStatus.Internal
 class ExistsOnProcessorImpl implements ExistsOnProcessor {
-    static final ClassNode ONLY_IN = makeWithoutCaching(OnlyIn)
-    static final ClassNode DIST = makeWithoutCaching(Dist)
+    static final ClassNode CLIENT_ONLY = makeWithoutCaching(ClientOnly)
+    static final ClassNode SERVER_ONLY = makeWithoutCaching(DedicatedServerOnly)
 
     @Override
     void process(AnnotatedNode node, List<Platform> platforms, ModuleNode ast, CompilePhase phase) {
-        List<Side> sides = platforms.findAll {it.loader === Loader.FORGE}.collect {it.side}.unique()
+        List<Side> sides = platforms.findAll {it.loader === Loader.QUILT}.collect {it.side}.unique()
         if (sides.size() === 1) {
             if (phase === CompilePhase.CANONICALIZATION) {
                 Side side = sides[0]
                 switch (side) {
                     case Side.CLIENT:
-                        node.addAnnotation(new AnnotationNode(ONLY_IN).tap {
-                            it.addMember('value', new PropertyExpression(
-                                    new ClassExpression(DIST),
-                                    new ConstantExpression(Dist.CLIENT)))
-                        })
+                        node.addAnnotation(new AnnotationNode(CLIENT_ONLY))
                         break
                     case Side.SERVER:
-                        node.addAnnotation(new AnnotationNode(ONLY_IN).tap {
-                            it.addMember('value', new PropertyExpression(
-                                    new ClassExpression(DIST),
-                                    new ConstantExpression(Dist.DEDICATED_SERVER)))
-                        })
+                        node.addAnnotation(new AnnotationNode(SERVER_ONLY))
                         break
                 }
             }
