--- Quilt/src/transform/resources/META-INF/services/io.github.groovymc.cgl.transform.environment.ExistsOnProcessor
@@ -1,1 +1,0 @@
-io.github.groovymc.cgl.quilt.transform.environment.ExistsOnProcessorImpl
